+++ plugins/core/hu.bme.mit.gamma.serializer.commandhandler/src/hu/bme/mit/gamma/serializer/commandhandler/ExecutionTraceMetricComputer.java
@@ -1,3 +1,13 @@
+/********************************************************************************
+ * Copyright (c) 2023 Contributors to the Gamma project
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * SPDX-License-Identifier: EPL-1.0
+ ********************************************************************************/
 package hu.bme.mit.gamma.serializer.commandhandler;
 
 import java.util.ArrayList;
@@ -25,7 +35,7 @@ import hu.bme.mit.gamma.util.GammaEcoreUtil;
 
 public class ExecutionTraceMetricComputer extends AbstractHandler {
 
-	protected Map<ExecutionTrace, Integer> traceMetrics = new HashMap<ExecutionTrace, Integer>();
+	protected final Map<ExecutionTrace, Integer> traceMetrics = new HashMap<ExecutionTrace, Integer>();
 	//
 	protected final FileUtil fileUtil = FileUtil.INSTANCE;
 	protected final GammaEcoreUtil ecoreUtil = GammaEcoreUtil.INSTANCE;