+++ plugins/core/hu.bme.mit.gamma.statechart.model/src/hu/bme/mit/gamma/statechart/derivedfeatures/StatechartModelDerivedFeatures.java
@@ -93,6 +93,7 @@ import hu.bme.mit.gamma.statechart.interface_.Port;
 import hu.bme.mit.gamma.statechart.interface_.RealizationMode;
 import hu.bme.mit.gamma.statechart.interface_.SimpleTrigger;
 import hu.bme.mit.gamma.statechart.interface_.TimeSpecification;
+import hu.bme.mit.gamma.statechart.interface_.TimeUnit;
 import hu.bme.mit.gamma.statechart.interface_.TopComponentArgumentsAnnotation;
 import hu.bme.mit.gamma.statechart.interface_.Trigger;
 import hu.bme.mit.gamma.statechart.interface_.UnfoldedPackageAnnotation;
@@ -571,7 +572,8 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 		Set<SynchronousComponent> types = new HashSet<SynchronousComponent>();
 		for (Component component : getAllComponents(parentPackage)) {
 			if (component instanceof SynchronousComponent) {
-				types.add((SynchronousComponent) component);
+				types.add(
+						(SynchronousComponent) component);
 			}
 		}
 		return types;
@@ -591,7 +593,8 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 		Set<StatechartDefinition> types = new HashSet<StatechartDefinition>();
 		for (Component component : getAllSynchronousComponents(parentPackage)) {
 			if (component instanceof StatechartDefinition) {
-				types.add((StatechartDefinition) component);
+				types.add(
+						(StatechartDefinition) component);
 			}
 		}
 		return types;
@@ -648,14 +651,16 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 			for (AsynchronousComponentInstance instance : asynchronousCompositeComponent.getComponents()) {
 				instances.add(instance);
 				AsynchronousComponent type = instance.getType();
-				instances.addAll(getAllInstances(type));
+				instances.addAll(
+						getAllInstances(type));
 			}
 		}
 		else if (component instanceof AsynchronousAdapter) {
 			AsynchronousAdapter asynchronousAdapter = (AsynchronousAdapter) component;
 			SynchronousComponentInstance wrappedComponent = asynchronousAdapter.getWrappedComponent();
 			instances.add(wrappedComponent);
-			instances.addAll(getAllInstances(wrappedComponent.getType()));
+			instances.addAll(
+					getAllInstances(wrappedComponent.getType()));
 		}
 		else if (component instanceof AbstractSynchronousCompositeComponent) {
 			AbstractSynchronousCompositeComponent synchronousCompositeComponent =
@@ -663,7 +668,8 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 			for (SynchronousComponentInstance instance : synchronousCompositeComponent.getComponents()) {
 				instances.add(instance);
 				SynchronousComponent type = instance.getType();
-				instances.addAll(getAllInstances(type));
+				instances.addAll(
+						getAllInstances(type));
 			}
 		}
 		return instances;
@@ -1206,7 +1212,7 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 	}
 	
 	public static List<ControlSpecification> getControlSpecifications(
-			AsynchronousAdapter adapter, Entry<Port, Event> portEvent) {
+			AsynchronousAdapter adapter, Object eventReference) {
 		List<ControlSpecification> controlSpecifications = new ArrayList<ControlSpecification>();
 		
 		for (ControlSpecification controlSpecification : adapter.getControlSpecifications()) {
@@ -1216,10 +1222,21 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 			}
 			if (trigger instanceof EventTrigger) {
 				EventTrigger eventTrigger = (EventTrigger) trigger;
-				EventReference eventReference = eventTrigger.getEventReference();
-				List<Entry<Port, Event>> inputEvents = getInputEvents(eventReference);
-				if (inputEvents.contains(portEvent)) {
-					controlSpecifications.add(controlSpecification);
+				EventReference controlEventReference = eventTrigger.getEventReference();
+				if (eventReference instanceof Entry<?, ?>) { // Port-event
+					List<Entry<Port, Event>> inputEvents = getInputEvents(controlEventReference);
+					@SuppressWarnings("unchecked")
+					Entry<Port, Event> portEvent = (Entry<Port, Event>) eventReference;
+					if (inputEvents.contains(portEvent)) {
+						controlSpecifications.add(controlSpecification);
+					}
+				}
+				else if (eventReference instanceof Clock clock) { // Clock
+					if (controlEventReference instanceof ClockTickReference clockTickReference) {
+						if (clockTickReference.getClock() == clock) {
+							controlSpecifications.add(controlSpecification);
+						}
+					}
 				}
 			}
 		}
@@ -1227,25 +1244,28 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 		return controlSpecifications;
 	}
 	
-	public static boolean isRunSpecification(AsynchronousAdapter adapter, Entry<Port, Event> portEvent) {
-		List<ControlSpecification> controlSpecifications = getControlSpecifications(adapter, portEvent);
+	public static boolean isRunSpecification(AsynchronousAdapter adapter, Object eventReference) {
+		List<ControlSpecification> controlSpecifications =
+				getControlSpecifications(adapter, eventReference);
 		
 		return controlSpecifications.stream()
 				.anyMatch(it -> it.getControlFunction() == ControlFunction.RUN_ONCE);
 	}
 	
-	public static boolean isComponentResetSpecification(AsynchronousAdapter adapter, Entry<Port, Event> portEvent) {
-		List<ControlSpecification> controlSpecifications = getControlSpecifications(adapter, portEvent);
+	public static boolean isComponentResetSpecification(AsynchronousAdapter adapter, Object eventReference) {
+		List<ControlSpecification> controlSpecifications =
+				getControlSpecifications(adapter, eventReference);
 		
 		return controlSpecifications.stream()
 				.anyMatch(it -> it.getControlFunction() == ControlFunction.RESET);
 	}
 	
 	public static Collection<? extends MessageQueue> getResetQueues(
-			AsynchronousAdapter adapter, Entry<Port, Event> portEvent) {
+			AsynchronousAdapter adapter, Object eventReference) {
 		Set<MessageQueue> resetMessageQueues = new LinkedHashSet<MessageQueue>();
 		
-		List<ControlSpecification> controlSpecifications = getControlSpecifications(adapter, portEvent);
+		List<ControlSpecification> controlSpecifications =
+				getControlSpecifications(adapter, eventReference);
 		List<MessageQueue> messageQueues = adapter.getMessageQueues();
 		
 		boolean resetQueue = controlSpecifications.stream()
@@ -1256,7 +1276,7 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 				.anyMatch(it -> it.getControlFunction() == ControlFunction.RESET_OTHER_MESSAGE_QUEUES);
 		
 		for (MessageQueue messageQueue : messageQueues) {
-			if (isStoredInMessageQueue(portEvent, messageQueue)) {
+			if (isStoredInMessageQueue(eventReference, messageQueue)) {
 				if (resetQueue) {
 					resetMessageQueues.add(messageQueue);
 				}
@@ -1297,6 +1317,19 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 		return queues;
 	}
 	
+	public static List<MessageQueue> getStoringMessageQueues(Clock clock) {
+		List<MessageQueue> queues = new ArrayList<MessageQueue>();
+		
+		AsynchronousAdapter adapter = ecoreUtil.getContainerOfType(clock, AsynchronousAdapter.class);
+		for (MessageQueue queue : adapter.getMessageQueues()) {
+			if (isStoredInMessageQueue(clock, queue)) {
+				queues.add(queue);
+			}
+		}
+		
+		return queues;
+	}
+	
 	public static boolean storesOnlyInternalEvents(MessageQueue queue) {
 		List<Entry<Port, Event>> storedEvents = getStoredEvents(queue);
 		return storedEvents.stream().allMatch(it -> isInternal(it.getKey()));
@@ -1317,6 +1350,14 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 					.collect(Collectors.toList()));
 		}
 		return List.copyOf(ports);
+	}
+	
+	public static List<Clock> getStoredClocks(MessageQueue queue) {
+		Collection<Clock> clocks = new LinkedHashSet<Clock>();
+		// To filter possible duplicates
+		clocks.addAll(
+				getClockTickReferences(queue).stream().map(it -> it.getClock()).toList());
+		return List.copyOf(clocks);
 	}
 	
 	public static List<Entry<Port, Event>> getStoredEvents(MessageQueue queue) {
@@ -1353,18 +1394,28 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 				.collect(Collectors.toList());
 	}
 	
-	public static EventPassing getEventPassing(MessageQueue queue,
-			Entry<Port, Event> portEvent) {
+	public static EventPassing getEventPassing(MessageQueue queue, Object eventReference) {
 		for (EventPassing eventPassing : queue.getEventPassings()) {
 			EventReference source = eventPassing.getSource();
 			if (source != null) { // Can be null due to reductions
-				List<Entry<Port, Event>> inputEvents = getInputEvents(source);
-				if (inputEvents.contains(portEvent)) {
-					return eventPassing;
+				// Entry<Port, Event>
+				if (eventReference instanceof Entry<?, ?> portEvent) {
+					List<Entry<Port, Event>> inputEvents = getInputEvents(source);
+					if (inputEvents.contains(portEvent)) {
+						return eventPassing;
+					}
+				}
+				// Clock
+				else if (eventReference instanceof Clock clock) {
+					if (source instanceof ClockTickReference clockTickReference) {
+						if (clockTickReference.getClock() == clock) {
+							return eventPassing;
+						}
+					}
 				}
 			}
 		}
-		throw new IllegalArgumentException("Not found event passing: " + portEvent);
+		throw new IllegalArgumentException("Not found event passing: " + eventReference);
 	}
 	
 	public static boolean isEventPassingCompatible(Port source, Port target) {
@@ -1400,9 +1451,8 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 		return isAcceptable;
 	}
 	
-	public static EventReference getTargetEventReference(MessageQueue queue,
-				Entry<Port, Event> portEvent) {
-		EventPassing eventPassing = getEventPassing(queue, portEvent);
+	public static EventReference getTargetEventReference(MessageQueue queue, Object eventReference) {
+		EventPassing eventPassing = getEventPassing(queue, eventReference);
 		EventReference target = eventPassing.getTarget();
 		if (target != null) {
 			return target;
@@ -1413,16 +1463,22 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 		}
 	}
 	
-	public static Entry<Port, Event> getTargetPortEvent(MessageQueue queue,
-			Entry<Port, Event> sourcePortEvent) {
-		Port sourcePort = sourcePortEvent.getKey();
-		Event sourceEvent = sourcePortEvent.getValue();
-		EventReference target = getTargetEventReference(queue, sourcePortEvent);
-		if (target instanceof AnyPortEventReference anyPortEventReference) {
+	public static Entry<Port, Event> getTargetPortEvent(MessageQueue queue, Object eventReference) {
+		EventReference target = getTargetEventReference(queue, eventReference);
+		if (target instanceof ClockTickReference) {
+			return null; // We cannot forward anything in this case
+		}
+		else if (target instanceof AnyPortEventReference anyPortEventReference) {
+			@SuppressWarnings("unchecked")
+			Entry<Port, Event> sourcePortEvent = (Entry<Port, Event>) eventReference;
+			Port sourcePort = sourcePortEvent.getKey();
+			Event sourceEvent = sourcePortEvent.getValue();
+			
 			Port targetPort = anyPortEventReference.getPort();
 			if (!isEventPassingCompatible(sourcePort, targetPort)) {
 				throw new IllegalArgumentException("Not the same interface: " + targetPort);
 			}
+			
 			List<Event> sourceEvents = getInputEvents(sourcePort);
 			int index = sourceEvents.indexOf(sourceEvent);
 			List<Entry<Port, Event>> targetEventIds = getInputEvents(target);
@@ -1433,7 +1489,7 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 			Event targetEvent = portEventReference.getEvent();
 			return new SimpleEntry<Port, Event>(port, targetEvent);
 		}
-		throw new IllegalArgumentException("Not known target: " + sourcePortEvent);
+		throw new IllegalArgumentException("Not known target: " + eventReference);
 	}
 	
 	public static List<EventReference> getSourceAndTargetEventReferences(MessageQueue queue) {
@@ -1447,6 +1503,29 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 			}
 		}
 		return eventReferences;
+	}
+	
+	public static List<ClockTickReference> getClockTickReferences(MessageQueue queue) {
+		List<ClockTickReference> eventReferences = new ArrayList<ClockTickReference>();
+		for (EventPassing eventPassing : queue.getEventPassings()) {
+			EventReference source = eventPassing.getSource();
+			if (source instanceof ClockTickReference clockTickReference) {
+				eventReferences.add(clockTickReference);
+			}
+		}
+		return eventReferences;
+	}
+	
+	public static int getEventId(MessageQueue queue, Clock clock) {
+		List<Entry<Port, Event>> storedEvents = getStoredEvents(queue);
+		int size = storedEvents.size();
+		
+		List<ClockTickReference> clockTickEventPassings = getClockTickReferences(queue);
+		ClockTickReference reference = javaUtil.getOnlyElement(
+				clockTickEventPassings.stream().filter(it -> it.getClock() == clock).toList());
+		int index = clockTickEventPassings.indexOf(reference) + 1; // Starts from event size + 1
+		
+		return size + index;
 	}
 	
 	public static int getEventId(MessageQueue queue, Entry<Port, Event> portEvent) {
@@ -1496,10 +1575,18 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 		return storedEvents.get(eventId - 1); // Starts from 1, 0 is the "empty cell"
 	}
 	
-	public static boolean isStoredInMessageQueue(
-			Entry<Port, Event> portEvent, MessageQueue queue) {
-		List<Entry<Port, Event>> storedEvents = getStoredEvents(queue);
-		return storedEvents.contains(portEvent);
+	public static boolean isStoredInMessageQueue(Object eventReference, MessageQueue queue) {
+		if (eventReference instanceof Entry<?, ?>) {
+			List<Entry<Port, Event>> storedEvents = getStoredEvents(queue);
+			@SuppressWarnings("unchecked")
+			Entry<Port, Event> portEvent = (Entry<Port, Event>) eventReference;
+			return storedEvents.contains(portEvent);
+		}
+		if (eventReference instanceof Clock clock) {
+			List<Clock> storedClocks = getStoredClocks(queue);
+			return storedClocks.contains(clock);
+		}
+		throw new IllegalArgumentException("Not known event reference: " + eventReference);
 	}
 	
 	public static boolean isTargetedInMessageQueue(
@@ -1598,11 +1685,13 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 	}
 	
 	public static List<Event> getInputEvents(Component component) {
-		return getInputEvents(getAllPorts(component));
+		return getInputEvents(
+				getAllPorts(component));
 	}
 	
 	public static List<Event> getOutputEvents(Component component) {
-		return getOutputEvents(getAllPorts(component));
+		return getOutputEvents(
+				getAllPorts(component));
 	}
 	
 	public static Port getBoundCompositePort(Port port) {
@@ -1889,7 +1978,9 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
     	else if (component instanceof AsynchronousAdapter) {
     		// Clocks maybe?
     		AsynchronousAdapter adapter = (AsynchronousAdapter) component;
-    		return isTimed(adapter.getWrappedComponent().getType());
+    		List<Clock> clocks = adapter.getClocks();
+    		SynchronousComponent type = adapter.getWrappedComponent().getType();
+			return isTimed(type) || !clocks.isEmpty();
     	}
     	else if (component instanceof AbstractAsynchronousCompositeComponent) {
     		AbstractAsynchronousCompositeComponent composite = (AbstractAsynchronousCompositeComponent) component;
@@ -2878,6 +2969,21 @@ public class StatechartModelDerivedFeatures extends ActionModelDerivedFeatures {
 			}
 		}
 		return time;
+	}
+	
+	public static Expression getTimeInMilliseconds(TimeSpecification timeout) {
+		Expression time = timeout.getValue();
+		Expression clonedTime = ecoreUtil.clone(time);
+		TimeUnit unit = timeout.getUnit();
+		switch (unit) {
+			case MILLISECOND:
+				return clonedTime;
+			case SECOND: {
+				return statechartUtil.wrapIntoMultiply(clonedTime, 1000);
+			}
+		default:
+			throw new IllegalArgumentException("Unexpected value: " + unit);
+		}
 	}
 	
 	public static Component getMonitoredComponent(StatechartDefinition adaptiveContract) {