+++ RELEASE-NOTES.md
@@ -245,3 +245,20 @@
 * Bugs related to orthogonal actions in the XSTS-UPPAAL transformation have been fixed.
 * Bugs related to region scheduling in the XSTS-UPPAAL transformation have been fixed.
 * Bugs related to command line arguments during verification have been fixed.
+
+## 2023.03.31. - Version 2.9.0
+
+### What is New
+
+* Gamma tasks now can be cancelled on the GUI.
+* The trace language (GTL) now supports variable declarations, too.
+* The composition language (GCL) now supports event redirection with regard to the target input port-event combination in message queues of asynchronous adapters, too.
+* The composition language (GCL) now supports message queue clearing control specifications of asynchronous adapters, too.
+* The statechart language (GSL) now supports the sequential, unordered and parallel execution of orthogonal regions.
+
+### Improvements
+
+* Model slicing algorithms for verification based on XSTS models has been improved.
+* Clock handling in XSTS-UPPAAL models has been improved.
+* The Promela mapping has been improved.
+* The model checking of asynchronous-reactive models has been improved.